fix: Calculator.java 내 사칙연산 메소드의 인자 타입을 double로 변경

+++ src/main/java/calculator/Calculator.java
@@ -2,18 +2,18 @@ package calculator;
 
 public class Calculator {
 
-    public double add(int a, int b) {
+    public double add(double a, double b) {
         return a + b;
     }
 
-    public double substract(int a, int b) {
+    public double substract(double a, double b) {
         return  a - b;
     }
-    public double multiply(int a, int b) {
+    public double multiply(double a, double b) {
         return a * b;
     }
 
-    public double divide(int a, int b) {
+    public double divide(double a, double b) {
         return  a / b;
     }
 }
